requirements.txt
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.6
entrypoints==0.3
flake8==3.7.9
idna==2.9
mccabe==0.6.1
mdutils==1.1.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyYAML==5.3.1
requests==2.23.0
typing==3.7.4.1
urllib3==1.25.9
